move daemon contract related logic to traits.
